requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings

python-dotenv

langchain-google-genai
PyMuPDF
python-docx
requests 
openai

pinecone-client
faiss-cpu 

langchain  
langchain-openai
numpy 